Working in different Files
We don't usually work in a single file while working with classes. We add a class in a separate file and just include it in our main.cpp file so that the users don't have to look at all that irrelevant material. We can do the process like described below:

1: Open visual Studio 2022, add a .cpp file.
2: In the Solution Explorer, write click on the name of the project.
3: Click on "add" and then on "Class".
You will get three files:
    1: .h file
    2: .cpp file
    3: main.cpp

open the file to visualize with an example.